chore: split handler cases in functions
